refactor: update ContentLayout import paths and add ContentWrapper component
