+++ src/app/courses/_components/content-layout.tsx
