+++ src/app/courses/_components/content-wrapper.tsx
