+++ src/app/courses/_components/news-and-events.tsx
