+++ src/app/courses/school-of-arts-and-social-sciences/ma-english/_components/core-faculty-members.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import FacultyAndStaff from "@/app/courses/miscellaneous/_components/faculty-and-staff";
+import FacultyAndStaff from "@/app/courses/_components/faculty-and-staff";
 
 const CoreFacultyMembers = () => {
 	return <FacultyAndStaff department="MA-ENGLISH" />;
