+++ src/app/courses/school-of-engineering/bsc-civil/_components/core-faculty-members.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-import FacultyAndStaff from "@/app/courses/miscellaneous/_components/faculty-and-staff";
+import FacultyAndStaff from "@/app/courses/_components/faculty-and-staff";
 
 const CoreFacultyMembers = () => {
 	return <FacultyAndStaff department="BSC-CIVIL" />;
